requirements.txt
fastapi
uvicorn
googlesearch-python
